+++ src/lib/infrastructure/persistence/persistence_container.dart
@@ -17,24 +17,24 @@ import 'package:whph/core/application/features/tasks/services/abstraction/i_task
 import 'package:whph/core/application/features/tasks/services/abstraction/i_task_time_record_repository.dart';
 import 'package:acore/acore.dart';
 import 'package:whph/core/application/features/app_usages/services/abstraction/i_app_usage_repository.dart';
-import 'package:whph/infrastructure/persistence/features/app_usages/drift_app_usage_ignore_rule_repository.dart';
-import 'package:whph/infrastructure/persistence/features/app_usages/drift_app_usage_repository.dart';
-import 'package:whph/infrastructure/persistence/features/app_usages/drift_app_usage_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/app_usages/drift_app_usage_tag_rule_repository.dart';
-import 'package:whph/infrastructure/persistence/features/app_usages/drift_app_usage_time_record_repository.dart';
-import 'package:whph/infrastructure/persistence/features/habits/drift_habit_records_repository.dart';
-import 'package:whph/infrastructure/persistence/features/habits/drift_habit_tags_repository.dart';
-import 'package:whph/infrastructure/persistence/features/habits/drift_habits_repository.dart';
-import 'package:whph/infrastructure/persistence/features/habits/drift_habit_time_record_repository.dart';
-import 'package:whph/infrastructure/persistence/features/notes/drift_note_repository.dart';
-import 'package:whph/infrastructure/persistence/features/notes/drift_note_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/settings/drift_settings_repository.dart';
-import 'package:whph/infrastructure/persistence/features/sync/drift_sync_device_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tags/drift_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tags/drift_tag_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_time_record_repository.dart';
+import 'package:whph/infrastructure/persistence/features/app_usages/repositories/drift_app_usage_ignore_rule_repository.dart';
+import 'package:whph/infrastructure/persistence/features/app_usages/repositories/drift_app_usage_repository.dart';
+import 'package:whph/infrastructure/persistence/features/app_usages/repositories/drift_app_usage_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/app_usages/repositories/drift_app_usage_tag_rule_repository.dart';
+import 'package:whph/infrastructure/persistence/features/app_usages/repositories/drift_app_usage_time_record_repository.dart';
+import 'package:whph/infrastructure/persistence/features/habits/repositories/drift_habit_records_repository.dart';
+import 'package:whph/infrastructure/persistence/features/habits/repositories/drift_habit_tags_repository.dart';
+import 'package:whph/infrastructure/persistence/features/habits/repositories/drift_habits_repository.dart';
+import 'package:whph/infrastructure/persistence/features/habits/repositories/drift_habit_time_record_repository.dart';
+import 'package:whph/infrastructure/persistence/features/notes/repositories/drift_note_repository.dart';
+import 'package:whph/infrastructure/persistence/features/notes/repositories/drift_note_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/settings/repositories/drift_settings_repository.dart';
+import 'package:whph/infrastructure/persistence/features/sync/repositories/drift_sync_device_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tags/repositories/drift_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tags/repositories/drift_tag_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/task_repository/drift_task_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/drift_task_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/drift_task_time_record_repository.dart';
 import 'package:whph/infrastructure/persistence/shared/contexts/drift/drift_app_context.dart';
 
 void registerPersistence(IContainer container) {
