+++ src/test/core/application/features/tasks/queries/task_queries_integration_test.dart
@@ -2,10 +2,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:drift/drift.dart' hide isNull, isNotNull;
 import 'package:whph/core/application/features/tasks/queries/get_task_query.dart';
 import 'package:whph/core/application/features/tasks/queries/get_list_tasks_query.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_time_record_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_tag_repository.dart';
-import 'package:whph/infrastructure/persistence/features/tags/drift_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/task_repository/drift_task_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/drift_task_time_record_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/drift_task_tag_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tags/repositories/drift_tag_repository.dart';
 import 'package:whph/infrastructure/persistence/shared/contexts/drift/drift_app_context.dart';
 import 'package:whph/core/domain/features/tasks/task.dart';
 import 'package:acore/acore.dart';
