+++ src/test/infrastructure/persistence/features/tasks/drift_task_repository_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:drift/drift.dart' hide isNull, isNotNull;
 import 'package:drift/native.dart';
 import 'package:whph/core/application/features/tasks/models/task_query_filter.dart';
-import 'package:whph/infrastructure/persistence/features/tasks/drift_task_repository.dart';
+import 'package:whph/infrastructure/persistence/features/tasks/repositories/task_repository/drift_task_repository.dart';
 import 'package:whph/infrastructure/persistence/shared/contexts/drift/drift_app_context.dart';
 import 'package:whph/core/domain/features/tasks/task.dart';
 import 'package:acore/acore.dart';
